Silence a warning.

+++ Plugins/CueSheet/CueSheetDecoder.m
@@ -176,7 +176,7 @@
 - (int)readAudio:(void *)buf frames:(UInt32)frames
 {
 	if (framePosition + frames > trackEnd) {
-		frames = trackEnd - framePosition;
+		frames = (UInt32)(trackEnd - framePosition);
 	}
 
 	if (!frames)
